Moved MeasurementType code to Measurement, fixed imports, removed measurement_type.py
